backend/requirements.txt
fastapi
uvicorn
openai
pydantic
pydantic-settings
PyYAML
databases[sqlite]
sqlalchemy
requests
lxml